Update Dockerfile to build from docker-taudem
Since changes to TauDEM are less frequent than changes to
rapid-watershed-delineation I have decided to encapsulate all the
instructions for building TauDEM in a separate repository.

This resolves an issue where bumping the version of RWD in this
Dockerfile would trigger an entire rebuild of TauDEM and all of its
dependencies. As a result of this change, updating the RWD docker image
and deploying new versions of RWD should be much quicker.

Connects #28